+++ src/app/oidc/device/page.tsx
@@ -10,8 +10,8 @@ import { normalizeUserCode } from "@/lib/oidc/device";
 import {
   buildDeviceFlowTargetLinkUri,
   issuerMatchesExpected,
-  thirdPartyInitiateSkipCookieName,
 } from "@/lib/oidc/third-party-initiate-login";
+import { thirdPartyInitiateSkipCookieName } from "@/lib/oidc/third-party-initiate-skip-cookie";
 import { getIssuer } from "@/lib/oidc/issuer-urls";
 
 type SearchParams = Record<string, string | string[] | undefined>;
