+++ src/lib/oidc/third-party-initiate-login.ts
@@ -1,8 +1,4 @@
-import { createHash } from "node:crypto";
-import { normalizeUserCode } from "@/lib/oidc/device";
 import { getPublicOrigin } from "@/lib/oidc/issuer-urls";
-
-const INITIATE_SKIP_MAX_AGE_SEC = 120;
 
 /** True only for loopback hostnames (cleartext HTTP allowed in non-production). */
 function isLocalhostHostname(hostname: string): boolean {
@@ -142,42 +138,4 @@ export function userCodeFromDeviceTargetLinkUri(
   } catch {
     return undefined;
   }
-}
-
-/**
- * HttpOnly cookie name: one skip flag per **client + device user code** so a
- * prior attempt (or logout / new CLI code) does not block third-party
- * `initiate_login_uri` redirects for unrelated flows for the same app.
- */
-export function thirdPartyInitiateSkipCookieName(
-  clientId: string,
-  userCode?: string | null,
-): string {
-  const normalized = userCode?.trim()
-    ? normalizeUserCode(userCode)
-    : "";
-  const h = createHash("sha256")
-    .update(`${clientId}\0${normalized}`)
-    .digest("hex")
-    .slice(0, 16);
-  return `pmth_tp_skip_${h}`;
-}
-
-export function initiateSkipCookieOptions(): {
-  httpOnly: boolean;
-  sameSite: "lax";
-  path: string;
-  maxAge: number;
-  secure: boolean;
-} {
-  const secure =
-    process.env.NODE_ENV === "production" ||
-    (process.env.NEXTAUTH_URL ?? "").startsWith("https:");
-  return {
-    httpOnly: true,
-    sameSite: "lax",
-    path: "/",
-    maxAge: INITIATE_SKIP_MAX_AGE_SEC,
-    secure,
-  };
 }
